make average score variable a float to be consistent with users table

+++ MySQL_Advanced/7-average_score.sql
@@ -6,11 +6,11 @@ CREATE PROCEDURE ComputeAverageScoreForUser(
     IN p_user_id INT
 )
 BEGIN
-    DECLARE v_average_score DECIMAL;
+    DECLARE v_average_score FLOAT;
 
     SELECT AVG(score) INTO v_average_score
     FROM corrections
-    WHERE user_id = p_user_id ;
+    WHERE user_id = p_user_id;
 
     UPDATE users
     SET average_score = v_average_score
